pyproject.toml
[project]
name = "mcp-clickhouse-documentation"
version = "0.1.0"
description = "MCP server for searching and reading ClickHouse documentation"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0,<3.0.0",
    "python-frontmatter>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mcp-clickhouse-documentation = "mcp_clickhouse_documentation.server:main"
clickhouse-docs-index = "mcp_clickhouse_documentation.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/mcp_clickhouse_documentation --cov-report=term-missing --cov-report=html"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]
